// src/shared/ComponentsUI/components/Typography/enums.ts
export enum Variant {
  h1 = 'h1',
  h2 = 'h2',
  h3 = 'h3',
  h4 = 'h4',
  h5 = 'h5',
  h6 = 'h6',
  text = 'p',
  smallText = 'small',
}

export enum FontStyle {
  italic = 'italic',
  normal = 'normal',
}

export enum FontWeight {
  black = 'black',
  bold = 'bold',
  extraBold = 'extra-bold',
  light = 'light',
  medium = 'medium',
  regular = 'regular',
  semiBold = 'semi-bold',
}